Add COSU functionality to TestDPC
Adding an EnableCosuActivity that is launched instead of
EnableProfileActivity when the admin extra contains a "cosu" tag. The
EnableCosuActivity downloads a config file and downloads, installs,
hides, enables app according to the config file. The user is then
dropped into one of three possible modes: default launcher, custom
launcher or single app mode.

Bug: 22648412
Change-Id: I2f272698cb3d4f786df4161a73a22ece46873b00