Merge sync-auth support back into non-sync-auth flow.
Remove the code duplication between sync-auth and other flows.
Originally it was intended that sync-auth would be styled differently
to the non-sync-auth flows for M, but it appears that both flows
ended being almost identical. We're starting to see issues due to
only one copy of code being updated, so lets address the root
problem.

Bug: 25988059
Change-Id: I8ed48dc8ee5081bcdc8b796e8e5ae2bf01c1e433