getContext should not be used as it is introduced in API 23
Override getContext function to return activtiy context to avoid similar
crash in the future. It works in our case as our fragments are always
added into an activtiy.

The proper fix should be using support v4 PreferenceFragment, but given
that we are going to release the testdpc soon, I prefer to have this temp
fix first.

Bug: 27491077
Change-Id: Ifebdd7217282d8c55207cd719dc1a970c374d41f