Add search functionality to testdpc
We have crawler to skim through all preference xmls of indexable
fragments and then we saved the index to a FTS table in sqlite database.
All searchable fragments extends a base class that except preference key
as argument, so that they will scroll down the target preference and
highlight it. Also, all searchable preference must implements isAvailable
in order to determine whether it should be shown in the search result page.

1. PreferenceCrawler is the crawler to skims through the XML
2. PolicySearchFragment process the search query and show the search result.
3. PreferenceIndexSqliteOpenHelper handles all sqlite operations

To add a new searachable fragment, you only need to
1. Make sure your fragment extends BaseSearchablePolicyPreferenceFragment
2. Register your fragments in IndexableFragments
And then you are good to go!

PS: the preference-v14 library is now broken, it is fixed in the internal
repo though. This commit should be merged only when the fix is in the
public build.

Fix: 26897552

Change-Id: I263db8594b3ec56f79ac0860269064732cb20865