Allow DO to set the organization name
In O, the DO will be allowed to set the organization name as well. The
current TestDPC UI for setting the organization name is buried deep
within PO settings. I prefer to leave it like that, because it fits
well there, and add a separate item for setting the organization name
as DO.

Bug: 32692748

Change-Id: I525da8f8bdf71572ae79db7721b4c9650a826a73