Fix isManagedProfileOwner() before profile is enabled.
Previously, we check the app is the profile owner and that there is more
than one profile to make sure it is a managed profile. However, before
the managed profile is enabled it does not count as a profile so the
test was giving incorrect results. Now, we assume it is a managed
profile if the app is a profile owner as TestDPC only supports being a
profile owner in a managed profile.

This only affects pre-M devices as N introduces DPM.isManagedProfile().

Bug: 62134588
Change-Id: I10bb0d7e6398e0da4e734609fa88fe28c3d8daed
(cherry picked from commit 384879e45dbdf719ae07fb18e74c336954c99a7c)