Implement ActivityOptions.setLockTaskMode().
In P, we introduce a way to launch a new task directly into LockTask
mode: ActivityOptions.setLockTaskMode(). This differs from previous
methods in that no special code is required in the target activity
itself (e.g. calling Activity.startLockTask() or setting the
lockTaskMode manifest attribute).

In this change, we implement this new API. Specifically, a new option is
added in PolicyManagementFragment to relaunch this activity with the
above ActivityOptions. This new option is put in parallel with the
existing "Start lock task mode" option.

Bug: 68687253
Test: Set TestDPC as DO, and then navigate to
      "TestDPC > Relaunch in lock task mode"
      Observe a new task is started with the current activity, and that
      this new task is put in LockTask mode directly. Also observe that
      the normal LockTask restrictions apply.

Change-Id: Ie0a511b2c5b5b3c1a38f58455de1b9a9a9e32d8b