Delegation-aware TestDPC
Mark some of the preferences available when TestDPC is not a DO/PO, but has the
appropriate delegation capabilities. Fix calls to DPM so TestDPC can call these
APIs as a delegated app (passing null as admin argument). Also make the
delegation fragment show the current delegations TestDPC has if it's no DO/PO.

To use this functionality, you should install the TestDPC-replica app alongside
the normal TestDPC as DO/PO, grant delegation to TestDPC-replica, and use it
to test the delegated functonalities.

Bug: 112982695
Test: manual
Change-Id: I630f532320f3ea37e37a3b3011a01e9611ca8f37