Get screen lock complexity and request specific complexity when setting new screen lock
For testing
- DevicePolicyManager.getPasswordComplexity
- DevicePolicyManager.EXTRA_PASSWORD_COMPLEXITY

Both require the permission GET_AND_REQUEST_SCREEN_LOCK_COMPLEXITY. It
is a normal permission which is granted on app install. This permission
is added to TestDpc in order to access the above API and intent extra.
This also means we cannot test the no permission case with this change.

Bug: 111173457
Change-Id: I01a0421fdc49e61734f3a72d2d4ed3070243bc18