Add UI for calling relinquishDevice
Add a TestDPC policy fragment for calling
DevicePolicyManager.relinquishDevice (a new method meant for a profile
owner of a corp-owned asset to remove the work profile and let the user
use the device as a personal device).

This is very similar to other policies - a key in the
device_policy_header and a corresponding handler in
PolicyManagementFragment.

Because this in an R API, attrs.xml and Util.java had to be changed to
recognize R APIs correctly.

Bug: 138709470
Test: Manual
Change-Id: Ic9b73be99cf6e2e58ceec70e0befbc1ef1c060d0