Call setCameraDisabled and getCameraDisabled on the parent profile
Previously, these APIs did not support explicitly querying the parent profile.
This CL allows the WP DPC to now call these methods, which means the WP DPC
can disable the camera device-wide and not just for the WP.

Added a new disable camera on parent toggle to TestDPC which can only be called
by a PO of an organization-owned device.

Bug: 138709470
Test: manual testing using Personal and WP TestDPC and Camera

Change-Id: Ifba278ed78b03b6fe29a85b8fa4128420dad7390