Optimization for creating new bundle in bundle array when the existing bundle contains multiple keys
Optimize for the case: create new bundle in Bundle array, if there is an existing bundle which contains multiple keys,
no need to re-create all the keys again in new created bundle, just automatically copy all keys from the existing one. This is extremely useful when the first bundle contains lots of keys.