Don't use onProfileProvisioningComplete on O+
Rely on ProvisioningSuccessActivity instead for O+. Currently, both
contain duplicate logic so the 'first one' can trigger first, but this
is unnecessary and causes problems since the receiver is a foreground
broadcast and setting the permission grant state can take too long.

We have to use the receiver version on <O although we haven't had
reports of problems.

Tried to put a unit test together but there are no tests for
DeviceAdminReceiver at the moment. This would be a negative test with no
corresponding positive test to depend on, and there was nothing
easy/obvious to add.

Test: manual, see doc in bug.
Bug: 177617306
Bug: 179259778
Change-Id: Id2451930969ede74c58f9efe364df71a987ef9da